modified notificationController,reportsController and models, routes and Services
