+++ backend/controllers/notificationController.js
@@ -1,5 +1,4 @@
 import Notification from '../models/Notification.js';
-import User from '../models/User.js';
 import * as NotificationService from '../services/notificationService.js';
 
 // Get all notifications for a user
@@ -204,72 +203,3 @@ export const sendBulkNotification = async (req, res) => {
 };
 
 
-// const Notification = require('../models/Notification');
-
-// // Get all notifications
-// exports.getAllNotifications = async (req, res) => {
-//   try {
-//     const notifications = await Notification.find()
-//       .sort({ createdAt: -1 });
-//     res.json(notifications);
-//   } catch (error) {
-//     res.status(500).json({ message: 'Error fetching notifications', error: error.message });
-//   }
-// };
-
-// // Get unread notifications
-// exports.getUnreadNotifications = async (req, res) => {
-//   try {
-//     const notifications = await Notification.find({ read: false })
-//       .sort({ createdAt: -1 });
-//     res.json(notifications);
-//   } catch (error) {
-//     res.status(500).json({ message: 'Error fetching unread notifications', error: error.message });
-//   }
-// };
-
-// // Mark notification as read
-// exports.markAsRead = async (req, res) => {
-//   try {
-//     const notification = await Notification.findByIdAndUpdate(
-//       req.params.id,
-//       { read: true },
-//       { new: true }
-//     );
-//     if (!notification) {
-//       return res.status(404).json({ message: 'Notification not found' });
-//     }
-//     res.json(notification);
-//   } catch (error) {
-//     res.status(500).json({ message: 'Error marking notification as read', error: error.message });
-//   }
-// };
-
-// // Delete notification
-// exports.deleteNotification = async (req, res) => {
-//   try {
-//     const notification = await Notification.findByIdAndDelete(req.params.id);
-//     if (!notification) {
-//       return res.status(404).json({ message: 'Notification not found' });
-//     }
-//     res.json({ message: 'Notification deleted successfully' });
-//   } catch (error) {
-//     res.status(500).json({ message: 'Error deleting notification', error: error.message });
-//   }
-// };
-
-// // Create new notification
-// exports.createNotification = async (req, res) => {
-//   try {
-//     const { title, message, type } = req.body;
-//     const notification = new Notification({
-//       title,
-//       message,
-//       type: type || 'info'
-//     });
-//     await notification.save();
-//     res.status(201).json(notification);
-//   } catch (error) {
-//     res.status(500).json({ message: 'Error creating notification', error: error.message });
-//   }
-// };
