+++ backend/models/Notification.js
@@ -1,34 +1,3 @@
-// const mongoose = require('mongoose');
-
-// const notificationSchema = new mongoose.Schema({
-//   title: {
-//     type: String,
-//     required: true,
-//     trim: true
-//   },
-//   message: {
-//     type: String,
-//     required: true,
-//     trim: true
-//   },
-//   type: {
-//     type: String,
-//     enum: ['info', 'warning', 'success'],
-//     default: 'info'
-//   },
-//   read: {
-//     type: Boolean,
-//     default: false
-//   },
-//   createdAt: {
-//     type: Date,
-//     default: Date.now
-//   }
-// });
-
-// module.exports = mongoose.model('Notification', notificationSchema); 
-
-
 import mongoose from 'mongoose';
 
 const notificationSchema = new mongoose.Schema({
